+++ borax/__init__.py
@@ -1,4 +1,4 @@
 # coding=utf8
 
-__version__ = '1.1.9'
+__version__ = '1.2.0b1'
 __author__ = 'kinegratii'